refactoring
https://online.sbis.ru/opendoc.html?guid=0db998a0-48b6-443d-b35e-3df569b522d6
